Moved files to types directory
