+++ src/app/public/index.ts
@@ -2,9 +2,9 @@ export * from './modules';
 
 // All future exports should be done in a single location,
 // to avoid circular references from barrel files.
-export * from './modules/date-range-picker/date-range-calculation';
-export * from './modules/date-range-picker/date-range-calculator-config';
-export * from './modules/date-range-picker/date-range-calculator-id';
-export * from './modules/date-range-picker/date-range-calculator-type';
+export * from './modules/date-range-picker/types/date-range-calculation';
+export * from './modules/date-range-picker/types/date-range-calculator-config';
+export * from './modules/date-range-picker/types/date-range-calculator-id';
+export * from './modules/date-range-picker/types/date-range-calculator-type';
 export * from './modules/date-range-picker/date-range-picker.module';
 export * from './modules/date-range-picker/date-range.service';
