+++ src/app/public/modules/date-range-picker/date-range-default-calculator-configs.ts
@@ -1,14 +1,14 @@
 import {
   SkyDateRangeCalculatorId
-} from './date-range-calculator-id';
+} from './types/date-range-calculator-id';
 
 import {
   SkyDateRangeCalculatorType
-} from './date-range-calculator-type';
+} from './types/date-range-calculator-type';
 
 import {
   SkyDateRangeDefaultCalculatorConfig
-} from './date-range-default-calculator-config';
+} from './types/date-range-default-calculator-config';
 
 import {
   SkyDateRangeRelativeValue
