+++ src/app/public/modules/date-range-picker/date-range-picker.component.html
@@ -40,6 +40,7 @@
     <sky-datepicker>
       <input
         formControlName="startDate"
+        [attr.aria-label]="startDateLabelResourceKey | skyLibResources"
         [attr.id]="dateRangePickerId + '-start-date'"
         [attr.placeholder]="dateFormat"
         [dateFormat]="dateFormat"
@@ -64,6 +65,7 @@
       <input
         formControlName="endDate"
         skyDatepickerInput
+        [attr.aria-label]="endDateLabelResourceKey | skyLibResources"
         [attr.id]="dateRangePickerId + '-end-date'"
         [attr.placeholder]="dateFormat"
         [dateFormat]="dateFormat"
