+++ src/app/public/modules/date-range-picker/date-range-picker.component.spec.ts
@@ -20,15 +20,15 @@ import {
 
 import {
   SkyDateRangeCalculation
-} from './date-range-calculation';
+} from './types/date-range-calculation';
 
 import {
   SkyDateRangeCalculatorId
-} from './date-range-calculator-id';
+} from './types/date-range-calculator-id';
 
 import {
   SkyDateRangeCalculatorType
-} from './date-range-calculator-type';
+} from './types/date-range-calculator-type';
 
 describe('Date range picker', function () {
   let fixture: ComponentFixture<DateRangePickerTestComponent>;
