+++ src/app/public/modules/date-range-picker/date-range-picker.component.ts
@@ -48,15 +48,15 @@ import {
 
 import {
   SkyDateRangeCalculation
-} from './date-range-calculation';
+} from './types/date-range-calculation';
 
 import {
   SkyDateRangeCalculatorId
-} from './date-range-calculator-id';
+} from './types/date-range-calculator-id';
 
 import {
   SkyDateRangeCalculatorType
-} from './date-range-calculator-type';
+} from './types/date-range-calculator-type';
 
 import {
   SkyDateRangeCalculator
