+++ src/app/public/modules/date-range-picker/fixtures/date-range-picker.component.fixture.ts
@@ -17,11 +17,11 @@ import {
 
 import {
   SkyDateRangeCalculation
-} from '../date-range-calculation';
+} from '../types/date-range-calculation';
 
 import {
   SkyDateRangeCalculatorId
-} from '../date-range-calculator-id';
+} from '../types/date-range-calculator-id';
 
 import {
   SkyDateRangePickerComponent
