+++ src/app/public/modules/date-range-picker/types/date-range-calculator-validate-function.ts
