+++ src/app/public/modules/datepicker/datepicker.component.spec.ts
@@ -168,7 +168,7 @@ describe('datepicker', () => {
 
     it('should apply aria-label to the datepicker input when none is provided', () => {
       fixture.detectChanges();
-      expect(nativeElement.querySelector('input').getAttribute('aria-label')).toBe('Date input field.');
+      expect(nativeElement.querySelector('input').getAttribute('aria-label')).toBe('Date input field');
     });
 
     it('should not overwrite aria-label on the datepicker input when one is provided', () => {
