+++ src/assets/locales/resources_en_US.json
@@ -1,7 +1,7 @@
 {
   "skyux_date_field_default_label": {
     "_description": "The label for the datepicker component's input field for screen readers when the consumer has not specified a label.",
-    "message": "Date input field."
+    "message": "Date input field"
   },
   "skyux_timepicker_button_label": {
     "_description": "Screen reader text for the timepicker component dropdown button",
@@ -13,7 +13,7 @@
   },
   "skyux_timepicker_input_default_label": {
     "_description": "The label for the timepicker component's input field for screen readers when the consumer has not specified a label.",
-    "message": "Time input field."
+    "message": "Time input field"
   },
   "skyux_date_range_picker_default_label": {
     "_description": "The default label for the date range picker.",
